pyproject.toml
[project]
name = "knowledge-graph-kb"
version = "0.1.0"
description = "A local-first knowledge graph knowledge base with OpenAI-powered ingestion and retrieval."
readme = "README.md"
authors = [
    { name = "2829798842", email = "2829798842@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.0",
    "fastapi>=0.116.1",
    "faiss-cpu>=1.13.2",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "openai>=1.108.1",
    "pydantic-settings>=2.11.0",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
]

[project.scripts]
knowledge-graph-kb = "main:cli"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.uv.build-backend]
module-name = "src"
module-root = ""

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
